+++ packages/common/src/visualizations/CartesianChartDataModel.ts
@@ -66,12 +66,30 @@ export class CartesianChartDataModel {
         this.type = args.type ?? ChartKind.VERTICAL_BAR;
     }
 
+    // Get the formatter for the tooltip, which has a simple callback signature
     static getTooltipFormatter(format: Format | undefined) {
         if (format === Format.PERCENT) {
             return (value: number) =>
                 applyCustomFormat(value, {
                     type: CustomFormatType.PERCENT,
                 });
+        }
+        return undefined;
+    }
+
+    // Get the formatter for the value label,
+    // which has more complex inputs
+    static getValueFormatter(format: Format | undefined) {
+        if (format === Format.PERCENT) {
+            // Echarts doesn't export the types for this function
+            return (params: any) => {
+                const value =
+                    params.value[params.dimensionNames[params.encode.y[0]]];
+
+                return applyCustomFormat(value, {
+                    type: CustomFormatType.PERCENT,
+                });
+            };
         }
         return undefined;
     }
@@ -470,7 +488,7 @@ export class CartesianChartDataModel {
         }
 
         const { type } = this;
-        const orgColors = colors;
+        const orgColors = colors ?? ECHARTS_DEFAULT_COLORS;
 
         const DEFAULT_X_AXIS_TYPE = VizIndexType.CATEGORY;
 
@@ -530,6 +548,11 @@ export class CartesianChartDataModel {
                         ? {
                               show: seriesValueLabelPosition !== 'hidden',
                               position: seriesValueLabelPosition,
+                              formatter: seriesFormat
+                                  ? CartesianChartDataModel.getValueFormatter(
+                                        seriesFormat,
+                                    )
+                                  : undefined,
                           }
                         : undefined,
                     labelLayout: {